compile banana

+++ src/voidsterdebugwindow.cpp
@@ -70,7 +70,7 @@ VoidsterdebugWindow::VoidsterdebugWindow(QWidget *parent, QString serialDev)
             //connect(store_ref, &store::batteryTemperatureChanged, )
             connect(store_ref, &store::hvChanged,InverterVoltage_Label, (void (FLabel::*)(short, short))&FLabel::setVisual );
             connect(store_ref, &store::bat_voltageChanged,Accumulator_Voltage_Label, (void (FLabel::*)(short, short))&FLabel::setVisual );
-            connect(store_ref, &store::max_cell_tempChanged,Max_cell_temp_label, (void (FLabel::*)(short, short))&FLabel::setVisual );
+            connect(store_ref, &store::max_cell_tempChanged,Max_cell_temp_label, (void (FLabel::*)(int, int))&FLabel::setVisual );
             //QObject::connect(store_ref,&store::menuChanged,this,&VoidsterdebugWindow::handleValueChanged);
 
             //connect(store_ref, &store::rpmChanged,VcuState_label,(void(FLabel::*)(short,short))&FLabel::setVisual);
